@@ -2,7 +2,7 @@
   <PropertyGroup>
     <TargetFramework>net6</TargetFramework>
     <RootNamespace>StarSwarm</RootNamespace>
-    <LangVersion>9.0</LangVersion>
+    <LangVersion>11.0</LangVersion>
     <Nullable>enable</Nullable>
     <RunAnalyzersDuringLiveAnalysis>True</RunAnalyzersDuringLiveAnalysis>
     <RunAnalyzersDuringBuild>True</RunAnalyzersDuringBuild>
