@@ -1,25 +1,21 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Godot;
+using System;
 
-namespace StarSwarm.Project.Autoload.AudioLibraries
+namespace StarSwarm.Project.Autoload.AudioLibraries;
+
+public partial class AudioLibrary : Node2D
 {
-    public partial class AudioLibrary : Node2D
+    private readonly RandomNumberGenerator _rng = new();
+
+    public Node GetAudioPlayer(String? nodePath = null)
     {
-        private readonly RandomNumberGenerator _rng = new RandomNumberGenerator();
-
-        public Node GetAudioPlayer(String? nodePath = null)
+        if (nodePath != null)
+            return GetNode<Node>(nodePath);
+        else
         {
-            if(nodePath != null)
-                return GetNode<Node>(nodePath);
-            else
-            {
-                _rng.Randomize();
+            _rng.Randomize();
 
-                return GetChild(_rng.RandiRange(0, GetChildCount() - 1));
-            }
+            return GetChild(_rng.RandiRange(0, GetChildCount() - 1));
         }
     }
 }
