@@ -1,33 +1,28 @@
-using Godot;
 using System;
 
-namespace StarSwarm.Project.GSAI_Framework.Behaviors
+namespace StarSwarm.Project.GSAI_Framework.Behaviors;
+
+public partial class GSAISteeringBehavior
 {
-    public partial class GSAISteeringBehavior
+    public Boolean IsEnabled = true;
+    public GSAISteeringAgent Agent { get; set; } = new GSAISteeringAgent();
+
+    public GSAISteeringBehavior(GSAISteeringAgent agent)
     {
-        public Boolean IsEnabled = true;
-        public GSAISteeringAgent Agent { get; set; } = new GSAISteeringAgent();
+        this.Agent = agent;
+    }
 
-        public GSAISteeringBehavior(GSAISteeringAgent agent)
+    public void CalculateSteering(GSAITargetAcceleration acceleration)
+    {
+        if (IsEnabled)
         {
-            this.Agent = agent;
+            _CalculateSteering(acceleration);
         }
-
-        public void CalculateSteering(GSAITargetAcceleration acceleration)
-        {
-            if(IsEnabled)
-            {
-                _CalculateSteering(acceleration);
-            }
-            else
-            {
-                acceleration.SetZero();
-            }
-        }
-
-        public virtual void _CalculateSteering(GSAITargetAcceleration acceleration)
+        else
         {
             acceleration.SetZero();
         }
     }
+
+    public virtual void _CalculateSteering(GSAITargetAcceleration acceleration) => acceleration.SetZero();
 }
