@@ -1,24 +1,21 @@
 using Godot;
-using System;
 
-namespace StarSwarm.Project.Ships.Enemies.SpaceCrab
+namespace StarSwarm.Project.Ships.Enemies.SpaceCrab;
+
+public partial class SpaceCrabStats : Node
 {
-    public partial class SpaceCrabStats : Node
-    {
-        // Declare member variables here. Examples:
-        // private int a = 2;
-        // private string b = "text";
+    // Declare member variables here. Examples:
+    // private int a = 2;
+    // private string b = "text";
 
-        // Called when the node enters the scene tree for the first time.
-        public override void _Ready()
-        {
-            
-        }
+    // Called when the node enters the scene tree for the first time.
+    public override void _Ready()
+    {
+    }
 
     //  // Called every frame. 'delta' is the elapsed time since the previous frame.
     //  public override void _Process(double delta)
     //  {
-    //      
+    //
     //  }
-    }
 }
