@@ -1,15 +1,8 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+namespace StarSwarm.Project.Ships.Player;
 
-namespace StarSwarm.Project.Ships.Player
+internal enum UpgradeChoices
 {
-    enum UpgradeChoices 
-    { 
-        HEALTH,
-        SPEED,
-        WEAPON 
-    }
+    HEALTH,
+    SPEED,
+    WEAPON
 }
