@@ -4,6 +4,7 @@ using System;
 public partial class HealthBarUpdater : Control
 {
     public TextureProgressBar HealthBar { get; set; } = default!;
+
     public override void _Ready()
     {
         HealthBar = GetNode<TextureProgressBar>("HealthBar");
@@ -25,8 +26,5 @@ public partial class HealthBarUpdater : Control
         HealthBar.Value = currentValue;
     }
 
-    public void OnMaxHealthUpdated(float maxHealth)
-    {
-        HealthBar.MaxValue = maxHealth;
-    }
+    public void OnMaxHealthUpdated(float maxHealth) => HealthBar.MaxValue = maxHealth;
 }
